requirements.txt
beautifulsoup4
django-sendgrid-webhook
stripe
django
requests
django-spurl
sendgrid
sentry-sdk
django-suit

python-dotenv
psycopg2
invoke

django-extensions
factory-boy
django-enumfield
django-anymail
django-storages[google]
gunicorn
django-credit-cards
django-filter
django-rest-knox

celery[redis]

django-maintenancemode

django-rest-framework

django_compressor
django-libsass
lxml
slimit
csscompressor

aiohttp[speedups]
